# BACK-END/API/src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Access Link API",
    "description": "API for Access",
    "version": "1.0.0"
  },
  "host": [
    "api.access-link.tech"
  ],
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format **Bearer &lt;token&gt;**"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/access/creerAccess": {
      "post": {
        "tags": [
          "Access"
        ],
        "description": "Créer un nouveau point d\\'accès",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Informations du point d'accès",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "macadress": {
                  "type": "string",
                  "example": "string"
                },
                "typePoint": {
                  "type": "string",
                  "example": "string"
                },
                "location": {
                  "type": "string",
                  "example": "string"
                },
                "nompoint": {
                  "type": "string",
                  "example": "string"
                },
                "active": {
                  "type": "string",
                  "example": "boolean"
                },
                "collabAutorise": {
                  "type": "string",
                  "example": "array"
                },
                "serviceAutorise": {
                  "type": "string",
                  "example": "array"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Point d'accès créé avec succès.",
            "schema": {
              "$ref": "#/definitions/Access"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/access/modifierAccess/{accessId}": {
      "put": {
        "tags": [
          "Access"
        ],
        "description": "Modifier un point d\\'accès existant",
        "parameters": [
          {
            "name": "accessId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID du point d'accès"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Nouvelles informations du point d'accès",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "macadress": {
                  "type": "string",
                  "example": "string"
                },
                "typePoint": {
                  "type": "string",
                  "example": "string"
                },
                "location": {
                  "type": "string",
                  "example": "string"
                },
                "nompoint": {
                  "type": "string",
                  "example": "string"
                },
                "active": {
                  "type": "string",
                  "example": "boolean"
                },
                "collabAutorise": {
                  "type": "string",
                  "example": "array"
                },
                "serviceAutorise": {
                  "type": "string",
                  "example": "array"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Point d'accès modifié avec succès.",
            "schema": {
              "$ref": "#/definitions/Access"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Point d'accès non trouvé."
          }
        }
      }
    },
    "/access/listeAccess": {
      "get": {
        "tags": [
          "Access"
        ],
        "description": "Obtenir la liste des points d\\'accès",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des points d'accès.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "$ref": "#/definitions/Access"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/access/pointsAccessibles": {
      "get": {
        "tags": [
          "Access"
        ],
        "description": "Obtenir les points d\\'accès accessibles par un collaborateur",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des points d'accès accessibles.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "$ref": "#/definitions/Access"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/access/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "description": "Obtenir les notifications pour le collaborateur connecté",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des notifications.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "$ref": "#/definitions/Notification"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/access/stats": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "description": "Obtenir les statistiques des points d\\'accès",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Statistiques des points d'accès.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "$ref": "#/definitions/Statistics"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/access/check/{token}/{macAdress}": {
      "get": {
        "tags": [
          "Access"
        ],
        "description": "Tester si un token a accès à un point",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Token du collaborateur"
          },
          {
            "name": "macAdress",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Adresse MAC du point d'accès"
          }
        ],
        "responses": {
          "200": {
            "description": "Résultat du test d'accès.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "hasAccess": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "boolean"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/access/config/{macAdress}": {
      "get": {
        "tags": [
          "Access"
        ],
        "description": "Récupérer la config du point d\\'accès",
        "parameters": [
          {
            "name": "macAdress",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Adresse MAC du point d'accès"
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration du point d'accès.",
            "schema": {
              "$ref": "#/definitions/AccessConfig"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Absence": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "collab": {
          "$ref": "#/definitions/Collaborateur"
        },
        "datedeb": {
          "type": "string",
          "example": "2023-07-15"
        },
        "periodeDeb": {
          "type": "number",
          "example": 1
        },
        "datefin": {
          "type": "string",
          "example": "2023-07-20"
        },
        "periodeFin": {
          "type": "number",
          "example": 2
        },
        "raison": {
          "type": "string",
          "example": "conge"
        },
        "description": {
          "type": "string",
          "example": "Congé annuel"
        },
        "accepte": {
          "type": "boolean",
          "example": true
        },
        "datereponse": {
          "type": "string",
          "example": "2023-07-16"
        },
        "dateDemande": {
          "type": "string",
          "example": "2023-07-01"
        },
        "reponseDe": {
          "$ref": "#/definitions/Collaborateur"
        },
        "modifierPar": {
          "$ref": "#/definitions/Collaborateur"
        }
      }
    },
    "Access": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "macadress": {
          "type": "string",
          "example": "00:14:22:01:23:45"
        },
        "typePoint": {
          "type": "string",
          "example": "pointeuse"
        },
        "location": {
          "type": "string",
          "example": "Bureau"
        },
        "nompoint": {
          "type": "string",
          "example": "Entrée principale"
        },
        "active": {
          "type": "boolean",
          "example": true
        },
        "collabAutorise": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Collaborateur"
          }
        },
        "serviceAutorise": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Service"
          }
        },
        "historique": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Historique"
          }
        }
      }
    },
    "Collaborateur": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "prenom": {
          "type": "string",
          "example": "John"
        },
        "nom": {
          "type": "string",
          "example": "Doe"
        },
        "mail": {
          "type": "string",
          "example": "john.doe@example.com"
        },
        "motdepasse": {
          "type": "string",
          "example": "password"
        },
        "fonction": {
          "type": "string",
          "example": "Développeur"
        },
        "grade": {
          "type": "string",
          "example": "collab"
        },
        "tokens": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Token"
          }
        },
        "absences": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Absence"
          }
        },
        "incident": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Incident"
          }
        },
        "historique": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Historique"
          }
        },
        "service": {
          "$ref": "#/definitions/Service"
        },
        "horairesdefault": {
          "$ref": "#/definitions/HorairesModele"
        },
        "horaire": {
          "$ref": "#/definitions/Horaire"
        },
        "actif": {
          "type": "boolean",
          "example": true
        },
        "valide": {
          "type": "boolean",
          "example": false
        },
        "date": {
          "type": "string",
          "example": "2023-07-01"
        }
      }
    },
    "Historique": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "date": {
          "type": "string",
          "example": "2023-07-15"
        },
        "collab": {
          "$ref": "#/definitions/Collaborateur"
        },
        "point": {
          "$ref": "#/definitions/Access"
        },
        "typeAction": {
          "type": "string",
          "example": "Entrée"
        },
        "actionAutorise": {
          "type": "boolean",
          "example": true
        },
        "statutUtilise": {
          "type": "string",
          "example": "Identité"
        }
      }
    },
    "Horaire": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "hDebLundi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinLundi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebMardi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinMardi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebMercredi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinMercredi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebJeudi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinJeudi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebVendredi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinVendredi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebSamedi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinSamedi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebDimanche": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinDimanche": {
          "type": "string",
          "example": "17:00:00"
        }
      }
    },
    "HorairesModele": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "nom": {
          "type": "string",
          "example": "Modèle standard"
        },
        "hDebLundi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinLundi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebMardi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinMardi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebMercredi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinMercredi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebJeudi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinJeudi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebVendredi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinVendredi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebSamedi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinSamedi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebDimanche": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinDimanche": {
          "type": "string",
          "example": "17:00:00"
        },
        "collabs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Collaborateur"
          }
        }
      }
    },
    "Incident": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "collab": {
          "$ref": "#/definitions/Collaborateur"
        },
        "description": {
          "type": "string",
          "example": "Problème de connexion"
        },
        "ouvert": {
          "type": "boolean",
          "example": true
        },
        "modifiePar": {
          "$ref": "#/definitions/Collaborateur"
        },
        "modifieLe": {
          "type": "string",
          "example": "2023-07-15"
        },
        "creeLe": {
          "type": "string",
          "example": "2023-07-10"
        },
        "creePar": {
          "type": "string",
          "example": "John Doe"
        },
        "dateIncident": {
          "type": "string",
          "example": "2023-07-09"
        },
        "reponse": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/IncidentReponse"
          }
        }
      }
    },
    "IncidentReponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "reponse": {
          "type": "string",
          "example": "Problème résolu"
        },
        "incident": {
          "$ref": "#/definitions/Incident"
        },
        "creeLe": {
          "type": "string",
          "example": "2023-07-11"
        }
      }
    },
...
    "Service": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "chefservice": {
          "$ref": "#/definitions/Collaborateur"
        },
        "nomservice": {
          "type": "string",
          "example": "Informatique"
        },
        "collabs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Collaborateur"
          }
        }
      }
    },
    "Token": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "collab": {
          "$ref": "#/definitions/Collaborateur"
        },
        "type": {
          "type": "string",
          "example": "auth"
        },
        "actif": {
          "type": "boolean",
          "example": true
        },
        "token": {
          "type": "string",
          "example": "token123"
        },
        "datecreation": {
          "type": "string",
          "example": "2023-07-15"
        }
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}